refactor(components): relocate components to new module structure
