+++ src/modules/common/sections/SearchPill/SearchPill.tsx
