+++ src/modules/map/sections/CollectionPointsMap.tsx
