+++ src/routes/index.tsx
@@ -11,9 +11,9 @@ import {
 import { For } from 'solid-js'
 
 import heroImage from '~/assets/hero-recycling.jpg'
-import { Navbar } from '~/components/Navbar'
 import { Button } from '~/components/ui/button'
 import { Card, CardContent } from '~/components/ui/card'
+import { Navbar } from '~/modules/common/sections/Navbar'
 
 const Home = () => {
   const stats = [
